+++ Changelog.md
@@ -13,6 +13,7 @@
 - Fix: Startup crash when Dave's Potioneering is installed
 - Fix: It's possible to scroll the Trophy System's list by clicking outside the list
 - Fix: The Universal Owner Changer cannot be put into a display case
+- Fix: The Trophy System's "Ignore Owner" option does not work
 
 --------------------------Changelog for v1.9.7 of SecurityCraft--------------------------